+++ packages/ui/src/hooks/usePaymentTokens.ts
@@ -1,11 +1,16 @@
 import { erc20ABI, useContractReads } from 'wagmi'
 import { fetchBalance } from 'wagmi/actions'
 import { Address, formatUnits, parseUnits, zeroAddress } from 'viem'
-import { useReservoirClient, useCurrencyConversions } from '.'
-import { useMemo } from 'react'
+import { useContext, useMemo } from 'react'
+import {
+  useReservoirClient,
+  useCurrencyConversions,
+  useSolverCapacities,
+} from '.'
 import { ReservoirChain } from '@reservoir0x/reservoir-sdk'
 import { PaymentToken } from '@reservoir0x/reservoir-sdk/src/utils/paymentTokens'
 import useSWR from 'swr'
+import { ProviderOptionsContext } from '../ReservoirKitProvider'
 
 export type EnhancedCurrency =
   | NonNullable<ReservoirChain['paymentTokens']>[0] & {
@@ -16,6 +21,8 @@ export type EnhancedCurrency =
       balance?: string | number | bigint
       currencyTotalRaw?: bigint
       currencyTotalFormatted?: string
+      maxItems?: number
+      maxPricePerItem?: bigint
     }
 
 const fetchNativeBalances = async (
@@ -45,26 +52,43 @@ export default function (
   preferredCurrencyTotalPrice: bigint,
   chainId?: number,
   nativeOnly?: boolean,
-  crossChainDisabled?: boolean
+  crossChainDisabled?: boolean,
+  listingCurrency?: EnhancedCurrency
 ) {
   const client = useReservoirClient()
+  const providerOptions = useContext(ProviderOptionsContext)
   const chain =
     chainId !== undefined
       ? client?.chains.find((chain) => chain.id === chainId)
       : client?.currentChain()
 
+  const includeListingCurrency =
+    providerOptions.alwaysIncludeListingCurrency !== false
+
   const allPaymentTokens = useMemo(() => {
     let paymentTokens = chain?.paymentTokens
 
-    if (nativeOnly) {
-      paymentTokens = paymentTokens?.filter(
-        (token) => token.address === zeroAddress
+    if (includeListingCurrency) {
+      const listingCurrencyAlreadyExists = paymentTokens?.some(
+        (token) =>
+          token.address.toLowerCase() ===
+            listingCurrency?.address?.toLowerCase() &&
+          token.chainId === listingCurrency?.chainId
       )
+      if (!listingCurrencyAlreadyExists && listingCurrency) {
+        paymentTokens?.push(listingCurrency)
+      }
     }
 
     if (crossChainDisabled) {
       paymentTokens = paymentTokens?.filter(
         (token) => token.chainId === chain?.id
+      )
+    }
+
+    if (nativeOnly) {
+      paymentTokens = paymentTokens?.filter(
+        (token) => token.address === zeroAddress
       )
     }
 
@@ -76,19 +100,42 @@ export default function (
       paymentTokens?.push(preferredCurrency)
     }
     return paymentTokens
-  }, [chain?.paymentTokens, preferredCurrency.address, nativeOnly])
+  }, [
+    chain?.paymentTokens,
+    preferredCurrency.address,
+    crossChainDisabled,
+    nativeOnly,
+    listingCurrency,
+    includeListingCurrency,
+  ])
 
   const nonNativeCurrencies = useMemo(() => {
     return allPaymentTokens?.filter(
       (currency) => currency.address !== zeroAddress
     )
-  }, [allPaymentTokens])
+  }, [
+    allPaymentTokens,
+    chain?.paymentTokens,
+    preferredCurrency.address,
+    crossChainDisabled,
+    nativeOnly,
+    listingCurrency,
+    includeListingCurrency,
+  ])
 
   const nativeCurrencies = useMemo(() => {
     return allPaymentTokens?.filter(
       (currency) => currency.address === zeroAddress
     )
-  }, [allPaymentTokens])
+  }, [
+    allPaymentTokens,
+    chain?.paymentTokens,
+    preferredCurrency.address,
+    crossChainDisabled,
+    nativeOnly,
+    listingCurrency,
+    includeListingCurrency,
+  ])
 
   const { data: nonNativeBalances } = useContractReads({
     contracts: open
@@ -110,6 +157,23 @@ export default function (
     {
       revalidateOnFocus: false,
     }
+  )
+
+  const crosschainChainIds = useMemo(() => {
+    if (crossChainDisabled) {
+      return []
+    } else {
+      return (
+        allPaymentTokens
+          ?.filter((token) => token?.chainId !== chain?.id)
+          ?.map((token) => token?.chainId) ?? []
+      )
+    }
+  }, [allPaymentTokens, crossChainDisabled])
+
+  const { data: solverCapacityChainIdMap } = useSolverCapacities(
+    open ? crosschainChainIds : [],
+    chain
   )
 
   const preferredCurrencyConversions = useCurrencyConversions(
@@ -180,6 +244,25 @@ export default function (
           ? formatUnits(usdTotalPriceRaw, 6)
           : undefined
 
+        let maxItems: EnhancedCurrency['maxItems'] = undefined
+        let maxPricePerItem: EnhancedCurrency['maxPricePerItem'] = undefined
+
+        if (
+          !crossChainDisabled &&
+          crosschainChainIds?.length > 0 &&
+          solverCapacityChainIdMap &&
+          currency.chainId !== chain?.id
+        ) {
+          const solverCapacity = solverCapacityChainIdMap.get(currency.chainId)
+
+          if (solverCapacity) {
+            maxItems = solverCapacity.maxItems
+            if (typeof solverCapacity.maxPricePerItem === 'string') {
+              maxPricePerItem = BigInt(solverCapacity.maxPricePerItem)
+            }
+          }
+        }
+
         return {
           ...currency,
           address: currency.address.toLowerCase(),
@@ -190,28 +273,60 @@ export default function (
           balance,
           currencyTotalRaw,
           currencyTotalFormatted,
+          maxItems,
+          maxPricePerItem,
         }
       })
-      .filter((currency) =>
-        currency.chainId !== chain?.id &&
-        (currency.currencyTotalRaw || 0) > 50000000000000000n
-          ? false
-          : true
-      )
       .sort((a, b) => {
-        // If user has a balance for the listed currency, return first. Otherwise sort currencies by total usdPrice
+        // If user has enough balance in the listing currency, return first. Otherwise sort currencies by balance and chainId
+
+        // User has enough balance in listing currency
         if (
-          a.address === preferredCurrency.address &&
-          a.chainId === preferredCurrency.chainId &&
-          Number(a.balance) > 0
-        )
+          listingCurrency &&
+          a.address === listingCurrency.address &&
+          a.chainId === listingCurrency.chainId &&
+          a.currencyTotalRaw &&
+          BigInt(a.balance) > a.currencyTotalRaw
+        ) {
           return -1
+        }
         if (
-          b.address === preferredCurrency.address &&
-          b.chainId === preferredCurrency.chainId &&
-          Number(b.balance) > 0
-        )
+          listingCurrency &&
+          b.address === listingCurrency.address &&
+          b.chainId === listingCurrency.chainId &&
+          b.currencyTotalRaw &&
+          BigInt(b.balance) > b.currencyTotalRaw
+        ) {
           return 1
+        }
+
+        // User has enough balance in non-listing currency
+        if (a.currencyTotalRaw && BigInt(a.balance) > a.currencyTotalRaw) {
+          return -1
+        }
+
+        if (b.currencyTotalRaw && BigInt(b.balance) > b.currencyTotalRaw) {
+          return 1
+        }
+
+        // Currency is the listing currency
+        if (
+          listingCurrency &&
+          a.address === listingCurrency.address &&
+          a.chainId === listingCurrency.chainId
+        ) {
+          return -1
+        }
+
+        if (
+          listingCurrency &&
+          b.address === listingCurrency.address &&
+          b.chainId === listingCurrency.chainId
+        ) {
+          return 1
+        }
+
+        // Otherwise sort by usdPrice and chaindId
         if (Number(b.usdPrice) === Number(a.usdPrice)) {
           if (
             a.chainId === preferredCurrency.chainId &&
@@ -237,6 +352,7 @@ export default function (
     allPaymentTokens,
     nonNativeBalances,
     nativeBalances,
+    listingCurrency,
   ])
 
   return paymentTokens