+++ packages/ui/src/modal/collect/mint/MintContent.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react'
+import React, { FC, useContext } from 'react'
 import { ChildrenProps, CollectStep } from '../CollectModalRenderer'
 import {
   Anchor,
@@ -24,7 +24,7 @@ import {
   faWallet,
 } from '@fortawesome/free-solid-svg-icons'
 import QuantitySelector from '../../QuantitySelector'
-import { CollectModalCopy } from '../CollectModal'
+import { CollectCallbackData, CollectModalCopy } from '../CollectModal'
 import { MintImages } from './MintImages'
 import { CollectCheckout } from '../CollectCheckout'
 import SigninStep from '../../SigninStep'
@@ -37,6 +37,7 @@ import { truncateAddress } from '../../../lib/truncate'
 import { SelectPaymentToken } from '../../SelectPaymentToken'
 import getChainBlockExplorerUrl from '../../../lib/getChainBlockExplorerUrl'
 import { CurrentStepTxHashes } from '../../CurrentStepTxHashes'
+import { ProviderOptionsContext } from '../../../ReservoirKitProvider'
 
 export const MintContent: FC<
   ChildrenProps & {
@@ -44,6 +45,7 @@ export const MintContent: FC<
     copy: typeof CollectModalCopy
     open: boolean
     setOpen: React.Dispatch<React.SetStateAction<boolean>>
+    onGoToToken?: (data: CollectCallbackData) => any
   }
 > = ({
   collection,
@@ -64,18 +66,22 @@ export const MintContent: FC<
   feeUsd,
   currentChain,
   contract,
+  address,
   hasEnoughCurrency,
   addFundsLink,
   disableJumperLink,
   transactionError,
   stepData,
   collectStep,
+  contentMode,
   setCollectStep,
   isConnected,
   collectTokens,
   copy,
   setOpen,
+  onGoToToken,
 }) => {
+  const providerOptions = useContext(ProviderOptionsContext)
   const pathMap = stepData?.path
     ? (stepData.path as Path[]).reduce(
         (paths: Record<string, Path>, path: Path) => {
@@ -88,6 +94,10 @@ export const MintContent: FC<
         {} as Record<string, Path>
       )
     : {}
+
+  const maxQuantity = paymentCurrency?.maxItems
+    ? paymentCurrency?.maxItems
+    : maxItemAmount
 
   const totalMints =
     stepData?.currentStep?.items?.reduce((total, item) => {
@@ -171,15 +181,15 @@ export const MintContent: FC<
                       ellipsify
                       css={{ width: '100%' }}
                     >
-                      {formatNumber(maxItemAmount)}{' '}
-                      {maxItemAmount > 1 ? 'items' : 'item'} available
+                      {formatNumber(maxQuantity)}{' '}
+                      {maxQuantity > 1 ? 'items' : 'item'} available
                     </Text>
                   </Flex>
                   <QuantitySelector
                     quantity={itemAmount}
                     setQuantity={setItemAmount}
                     min={1}
-                    max={maxItemAmount}
+                    max={maxQuantity}
                     css={{
                       width: '100%',
                       justifyContent: 'space-between',
@@ -257,20 +267,43 @@ export const MintContent: FC<
               >
                 <Text style="h6">You Pay</Text>
                 <Flex direction="column" align="end" css={{ gap: '$1' }}>
-                  <FormatCryptoCurrency
-                    chainId={chainId}
-                    textStyle="h6"
-                    amount={paymentCurrency?.currencyTotalRaw}
-                    address={paymentCurrency?.address}
-                    decimals={paymentCurrency?.decimals}
-                    symbol={paymentCurrency?.name}
-                    logoWidth={18}
-                  />
-                  <FormatCurrency
-                    amount={paymentCurrency?.usdTotalPriceRaw}
-                    style="tiny"
-                    color="subtle"
-                  />
+                  {providerOptions.preferDisplayFiatTotal ? (
+                    <>
+                      <FormatCurrency
+                        amount={paymentCurrency?.usdTotalPriceRaw}
+                        style="h6"
+                        color="base"
+                      />
+                      <FormatCryptoCurrency
+                        chainId={chainId}
+                        textStyle="tiny"
+                        textColor="subtle"
+                        amount={paymentCurrency?.currencyTotalRaw}
+                        address={paymentCurrency?.address}
+                        decimals={paymentCurrency?.decimals}
+                        symbol={paymentCurrency?.symbol}
+                        logoWidth={12}
+                      />
+                    </>
+                  ) : (
+                    <>
+                      <FormatCryptoCurrency
+                        chainId={chainId}
+                        textStyle="h6"
+                        textColor="base"
+                        amount={paymentCurrency?.currencyTotalRaw}
+                        address={paymentCurrency?.address}
+                        decimals={paymentCurrency?.decimals}
+                        symbol={paymentCurrency?.symbol}
+                        logoWidth={18}
+                      />
+                      <FormatCurrency
+                        amount={paymentCurrency?.usdTotalPriceRaw}
+                        style="tiny"
+                        color="subtle"
+                      />
+                    </>
+                  )}
                 </Flex>
               </Flex>
             </Flex>
@@ -351,6 +384,7 @@ export const MintContent: FC<
             currency={paymentCurrency}
             setCurrency={setPaymentCurrency}
             goBack={() => setCollectStep(CollectStep.Idle)}
+            itemAmount={itemAmount}
           />
         </Flex>
       )}
@@ -561,7 +595,10 @@ export const MintContent: FC<
             <Text style="h5" css={{ px: '$5' }}>
               Your mint is complete!
             </Text>
-            <MintImages stepData={stepData} contract={contract} />
+            <MintImages
+              stepData={stepData}
+              tokenKind={collection?.contractKind}
+            />
             <Flex align="center" css={{ gap: '$2', px: '$5' }}>
               <Box
                 css={{
@@ -626,9 +663,39 @@ export const MintContent: FC<
             </Flex>
           </Flex>
           <Flex css={{ width: '100%', px: '$4' }}>
-            <Button css={{ width: '100%' }} onClick={() => setOpen(false)}>
-              {copy.mintCtaClose}
-            </Button>
+            {!!onGoToToken ? (
+              <>
+                <Button
+                  onClick={() => {
+                    setOpen(false)
+                  }}
+                  css={{ flex: 1 }}
+                  color="ghost"
+                >
+                  {copy.mintCtaClose}
+                </Button>
+                <Button
+                  style={{ flex: 1 }}
+                  color="primary"
+                  onClick={() => {
+                    onGoToToken({
+                      collectionId: collection?.id,
+                      maker: address,
+                      stepData,
+                      contentMode,
+                    })
+                  }}
+                >
+                  {copy.mintCtaGoToToken.length > 0
+                    ? copy.mintCtaGoToToken
+                    : `View ${successfulMints > 1 ? 'Tokens' : 'Token'}`}
+                </Button>
+              </>
+            ) : (
+              <Button css={{ width: '100%' }} onClick={() => setOpen(false)}>
+                {copy.mintCtaClose}
+              </Button>
+            )}
           </Flex>
         </Flex>
       )}