+++ packages/ui/src/modal/collect/sweep/SweepContent.tsx
@@ -1,4 +1,4 @@
-import React, { FC } from 'react'
+import React, { FC, useContext } from 'react'
 import { ChildrenProps, CollectStep } from '../CollectModalRenderer'
 import {
   Anchor,
@@ -36,6 +36,7 @@ import { formatNumber } from '../../../lib/numbers'
 import { truncateAddress } from '../../../lib/truncate'
 import getChainBlockExplorerUrl from '../../../lib/getChainBlockExplorerUrl'
 import { CurrentStepTxHashes } from '../../CurrentStepTxHashes'
+import { ProviderOptionsContext } from '../../../ReservoirKitProvider'
 
 export const SweepContent: FC<
   ChildrenProps & {
@@ -76,6 +77,7 @@ export const SweepContent: FC<
   copy,
   setOpen,
 }) => {
+  const providerOptions = useContext(ProviderOptionsContext)
   const hasTokens = orders && orders.length > 0
 
   const is1155 = collection?.contractKind === 'erc1155'
@@ -91,6 +93,10 @@ export const SweepContent: FC<
     mostExpensiveToken?.currency?.toLowerCase() != paymentCurrency?.address
       ? mostExpensiveToken?.buyInQuote
       : mostExpensiveToken?.totalPrice
+
+  const maxQuantity = paymentCurrency?.maxItems
+    ? paymentCurrency?.maxItems
+    : maxItemAmount
 
   const pathMap = stepData?.path
     ? (stepData.path as Path[]).reduce(
@@ -184,15 +190,15 @@ export const SweepContent: FC<
                     ellipsify
                     css={{ width: '100%' }}
                   >
-                    {formatNumber(maxItemAmount)}{' '}
-                    {maxItemAmount === 1 ? 'item' : 'items'} available
+                    {formatNumber(maxQuantity)}{' '}
+                    {maxQuantity === 1 ? 'item' : 'items'} available
                   </Text>
                 </Flex>
                 <QuantitySelector
                   quantity={itemAmount}
                   setQuantity={setItemAmount}
                   min={1}
-                  max={maxItemAmount}
+                  max={maxQuantity}
                   css={{
                     width: '100%',
                     justifyContent: 'space-between',
@@ -314,20 +320,43 @@ export const SweepContent: FC<
             >
               <Text style="h6">You Pay</Text>
               <Flex direction="column" align="end" css={{ gap: '$1' }}>
-                <FormatCryptoCurrency
-                  chainId={chainId}
-                  textStyle="h6"
-                  amount={paymentCurrency?.currencyTotalRaw}
-                  address={paymentCurrency?.address}
-                  decimals={paymentCurrency?.decimals}
-                  symbol={paymentCurrency?.symbol}
-                  logoWidth={18}
-                />
-                <FormatCurrency
-                  amount={paymentCurrency?.usdTotalPriceRaw}
-                  style="tiny"
-                  color="subtle"
-                />
+                {providerOptions.preferDisplayFiatTotal ? (
+                  <>
+                    <FormatCurrency
+                      amount={paymentCurrency?.usdTotalPriceRaw}
+                      style="h6"
+                      color="base"
+                    />
+                    <FormatCryptoCurrency
+                      chainId={chainId}
+                      textStyle="tiny"
+                      textColor="subtle"
+                      amount={paymentCurrency?.currencyTotalRaw}
+                      address={paymentCurrency?.address}
+                      decimals={paymentCurrency?.decimals}
+                      symbol={paymentCurrency?.symbol}
+                      logoWidth={12}
+                    />
+                  </>
+                ) : (
+                  <>
+                    <FormatCryptoCurrency
+                      chainId={chainId}
+                      textStyle="h6"
+                      textColor="base"
+                      amount={paymentCurrency?.currencyTotalRaw}
+                      address={paymentCurrency?.address}
+                      decimals={paymentCurrency?.decimals}
+                      symbol={paymentCurrency?.symbol}
+                      logoWidth={18}
+                    />
+                    <FormatCurrency
+                      amount={paymentCurrency?.usdTotalPriceRaw}
+                      style="tiny"
+                      color="subtle"
+                    />
+                  </>
+                )}
               </Flex>
             </Flex>
           </Flex>
@@ -414,6 +443,7 @@ export const SweepContent: FC<
             currency={paymentCurrency}
             setCurrency={setPaymentCurrency}
             goBack={() => setCollectStep(CollectStep.Idle)}
+            itemAmount={itemAmount}
           />
         </Flex>
       )}