+++ packages/ui/src/modal/editListing/EditListingModal.tsx
@@ -338,11 +338,13 @@ export function EditListingModal({
                         }
                       }}
                     >
-                      {expirationOptions.map((option) => (
-                        <Select.Item key={option.text} value={option.value}>
-                          <Select.ItemText>{option.text}</Select.ItemText>
-                        </Select.Item>
-                      ))}
+                      {expirationOptions
+                        .filter(({ value }) => value !== 'custom')
+                        .map((option) => (
+                          <Select.Item key={option.text} value={option.value}>
+                            <Select.ItemText>{option.text}</Select.ItemText>
+                          </Select.Item>
+                        ))}
                     </Select>
                   </Box>
                   <Flex